Connected API to daily, monthly and yearly views
