--- habit-tracker-app/src/calendarviews/calendarviewsApp.css
@@ -1,5 +1,0 @@
-.calendarviews-app {
-    color:#fc7a57;
-	border-style: solid;
-	text-align: center;
-}
