+++ habit-tracker-app/src/pages/habit-tracker.jsx
@@ -11,7 +11,6 @@ import PlantsApp from "../plants/plantsApp.jsx";
 import SocialisingApp from "../socialising/socialisingApp.jsx";
 import SocialmediaApp from "../socialmedia/socialmediaApp.jsx";
 import Grid from "../components/Grid";
-import CalendarviewsApp from "../calendarviews/calendarviewsApp.jsx";
 import SocialMediaSharing from "../socialmediasharing/socialmediasharing.jsx";
 import HydrationApp2 from "../components/HydrationApp2.jsx";
 
@@ -33,7 +32,6 @@ function HabitTracker() {
       <SocialisingApp />
       <SocialmediaApp />
       <HydrationApp2 />
-      <CalendarviewsApp />
     </div>
     <div className="socialmedia">
       <SocialMediaSharing />
